minor fixes with regard to the vegetarianMeal POST request

+++ server/src/main/java/server/group12/VegetarianMealUsageRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import java.util.List;
 
 public interface VegetarianMealUsageRepository extends MongoRepository<VegetarianMealEntry, String> {
-    public VegetarianMealEntry getByUserId(long userId);
+    public VegetarianMealEntry getByUserId(int userId);
     public List<VegetarianMealEntry> getByVegetarianMeal(String vegetarianMeal, int points, int co2);
 }
